requirements.txt
Flask==2.3.3
gunicorn==21.2.0
pyrebase4==4.7.1
python-dotenv==1.0.0
requests==2.31.0